[r=blake-rouse][bug=][author=blake-rouse] Add the ability to double click a disabled element to enable it.
By default browsers do not register click events for disabled elements, but this overlays the transcluded element with an overlay allowing click events to still be registered. Used on the node details page enter editing mode for the summary and power sections. Give a nice little shortcut instead of having to click edit.