[r=andreserl][bug=][author=blake-rouse] Adds device_actions to the GeneralHandler and GeneralManager to retrieve actions that can only be performed on devices.
Changed how actions are computed for the NodeAction class. Added installable_only property allowing a list of actions to be built only for devices, without passing a node to all the classes. Adjusted how permissions are calculated as they are a little different for devices thane from nodes. Renamed actions to node_actions, to show the difference between actions that can be completed on nodes vs those completed on devices.