[r=blake-rouse][bug=][author=allenap] Make it possible to cancel a failing start-up.
Previously start-up would sleep in its thread if it failed. This thread was not the main thread so it would not receive interrupt events, and would therefore KeyboardInterrupt would not be called within it. Retries are now managed in the reactor, which can cancel outstanding tasks when interrupts are received.