[r=rvb][bug=1453954][author=blake-rouse] Don't power query a node in the node power monitor service if a power action is on going.
This fixes two issues in one go. The first is bug 1453945, which can cause a releasing node to go to ready while the power action is still occurring causing a following call to start the node to fail as another power action is already in progress. The second removes the issue where some BMC's do not allow multiple connections. This doesn't fix the issue where the same BMC is being used for different machines.