[r=blake-rouse][bug=1436541][author=mpontillo] Remove obsolete 'Pending' cluster status. New statuses are "Enabled" and "Disabled". (previously: "Accepted" and "Rejected".)
This change includes two migrations:

 - One to "un-confuse" South, since it didn't fully understand that DomainNameField isn't somehow special
 - A second to do a data migration to remove any clusters with PENDING status (and change them to DISABLED)

The remainder of the commit simplifies the UI, adds the "Status" column, removes the cluster status links in the heading, and simplifies the UI overall. In doing so, many obsolete test cases could also be removed, related to:

 - The "Accept All" and "Reject All" buttons (and related back-end functionality) were removed
 - The filter URLs have been removed (to view only "pending" or "rejected" clusters
 - The greatly simplified view code

It was also necessary to add a 'map_enum_unique_values', because 'map_enum' was including the "DEFAULT" entry in the enum, leading to duplicates that were now confusing test cases. (since the DEFAULT changed from PENDING to ENABLED.)