[r=blake-rouse][bug=1457786,1457921][author=mpontillo] Set the boot order to  (network, hd) during probe-and-enlist for virsh.
This is accomplished by using "virsh define" to redefine the configuration for the virtual machine.

Note that a small possibility exists for a race condition, if the VM configuration is changed after we get the XML, but before we modify it and push the changes. Unfortunately, we'll have to live with this possibility, due to the limitations of virsh.