[r=rvb][bug=][author=blake-rouse] Add storage_layout option to the acquire API. Add set_storage_layout to the NodeHandler.
Adding storage_layout to the acquire API allows a user to set the initial storage_layout when the node is acquired. The set_storage_layout call allows a user with an allocated node to change the storage layout. This will overwrite their current storage layout, this is noted in the docstring of the method.