[r=mpontillo][bug=][author=blake-rouse] Force the partition table type based on the boot disk and the bios_boot_method.
When a node is booted using UEFI its boot disk will always use a GPT partition table and have an EFI partition. If not uefi the boot disk will always have a MBR partition and no EFI partition. Any block device that is not the boot block device will be by default GPT.