[r=blake-rouse][bug=][author=allenap] Fix upgrades involving migration 0139 from lp:maas/1.8 to trunk.
Migration 0139 in lp:maas/1.8 is a backport of migration 0150 from trunk. Unfortunately this causes all sorts of issues when a user of 1.8  upgrades to 1.9 or later.  Fixing this involved pulling 0139 into trunk, adjusting the name of the current 0139 to preserve ordering, then removing migration 0150. Users of trunk will unfortunately be faced with problems migrating, but users of 1.8 should be spared any unpleasantness.