[r=blake-rouse][bug=][author=mpontillo] Move claim_static_ips() function from MacAddress model object to Interface model object.
This change means that a MacAddress object MUST have a related Interface object, otherwise a user will not be able to use that MAC to claim static IP addresses. Therefore, the API code has been changed so that if a user claims an arbitrary MAC, a related interface will automatically be created.