[r=andreserl][bug=][author=blake-rouse] Make the storage configuration persistent instead of being cleared and reset on each deployment.
When a node is acquired all user mountable filesystems will be copied and acquired=True will be set. When the node is acquired on those filesystems are modified the original filesystems remain. When the node is released the created filesystems will be deleted. This make it where the storage configuration does not once the node is acquired for that deployment. It also means that only formatting and mounting can be done once a node is acquired.