[r=blake-rouse][bug=1505029][author=mpontillo] When registering a cluster interface, try to only add physical Ethernet interfaces automatically.
Due to the possibility that the MAAS cluster daemon is deployed on a container (and thus only has virtual interfaces), the following heuristic is used:

(1) Determine which Ethernet interfaces can be shown to be backed by hardware. (these will be associated with a device driver module)

(1a) Eliminate wireless cards from the set of interfaces found in (1)

(2) If no interfaces are found after step (1), fall back to interface naming conventions. A typical container, for example, will only have interfaces with names starting with 'eth'.

The need for this commit stems from the "fabric explosion" problem, where undesired interfaces (such as bridges or WLAN interfaces) may register with the region, which otherwise could automatically (and incorrectly) create excessive fabrics.