[r=mpontillo,allenap][bug=][author=blake-rouse] Support django version 1.8 and new piston3. Add new dbupgrade command.
This allows MAAS to run under django 1.8 which will the the version we will use for Xenial and on. Changes from piston to piston3. This branch does not address running under python3, but it uses piston3 which supports both python2.7 and python3. Add new "dbupgrade" command that allows MAAS to upgrade from any previous version of MAAS to version of MAAS running under django 1.8.