[r=andreserl][bug=][author=andreserl] Allow BMC to be on a network where a rack controller does not have direct access.
When a node is to be powered on or off and no direct rack controllers can be identified the region will ask all connected rack controllers to figure out which rack controllers can access that BMC by performing a power query. A many to many relationship between the rack controllers and BMC is updated to mark which rack controllers are routeable and which are not.