[r=allenap][bug=][author=mpontillo] Add utility scripts to create an LXD profile, give access to the MAAS sandbox, create an LXD image, and run tests in an isolated environment.
For convenience, two make targets have been added:

    make lxd - Creates the LXD image that will be used for testing

    make test+lxd - Runs the test suite on the LXD image

To use these scripts manually:

(1) Run utilities/configure-lxd-profile

This will create a LXD profile called "maas-dev" which will be used to share your MAAS sandboxes with the container, and create the container with a lax security policy (since these are only to be used for development).

(2) Run utilities/create-lxd-xenial-image

This will download the Xenial cloud image (if needed), then run "make install-dependencies" within the container, and locally publish an image. (so that this large download only has to happen once)

(3) Run utilities/isolated-make-test

This will create an ephemeral container using the image created in step (2) and use that container to run the test suite.