[r=blake-rouse][bug=][author=mpontillo] Add models for mDNS and neighbour (ARP) data. Enhance interface model to support the configuration of network discovery.
The Neighbour model adds a table to track which [ARP] neighbours were observed on each rack interface.

The MDNS model adds a table to track observed mDNS records.

Methods have been added to the `Interface` class which allows the records to be updated after the region locates the Interface for which the data is being reported.

Enhancements have been made to the Interface model so that mDNS discovery, neighbour discovery, and active discovery can be enabled and disabled at a granular level by the region. In particular, the following new fields have been added to the Interface model:
 * neighbour_discovery_state - determines whether or not to passively look for neighbour discovery (ARP) packets on the interface.
 * mdns_discovery_state - determines whether or not to take note of hostnames found on the interface using `avahi-browse`.
 * active_discovery_state - determines whether or not actively scanning for discovery purposes is acceptable on this interface.
 * active_discovery_params - parameters for active network discovery, to be sent to the rack controller if active discovery needs to be reconfigured.

Please note that the aforementioned changes to the Interface model are NOT intended to be user-editable. These values are necessary because we want to be able to compute the discovery state [on any region controller] and store it in the database, where it can be seen by any other region controller. Then we can propagate it to the rack controller(s) when a trigger notices the changes. Storing the expected state of network discovery in the most granular way possible (at the interface level) allows us the most flexibility in the future, when we could decide we want to modify discovery settings globally, per-VLAN, per-fabric, or per-interface.

Finally, added a drive-by fix to allow the check-imports script to properly handle statements such as "from . import <whatever>". (I experimented with using that syntax in this branch, then ultimately decided not to.)