[r=newell-jensen][bug=][author=newell-jensen] Adds SSH key source importing functionality to the API.  User can import SSH keys from launchpad or github.
For example, if the user's login sessions is named dev they could do something like this:

$ maas dev sshkeys import protocol=lp auth_id=user-id

where user-id is their launchpad id. Similarly, they could also do the same thing with github:

$ maas dev sshkeys import protocol=gh auth_id=user-id

If the protocol is left out, it will choose `lp` by default:

$ maas dev sshkeys import auth_id=user-id

The above actions will import the SSH keys from the respective accounts.

In a future branch we will most likely be changing the maascli in another branch so that we can do something similar to this:

$ maas dev sshkeys import lp:user-id

or

$ maas dev sshkeys import gh:user-id

with a default of

$ maas dev sshkeys import user-id

choosing `lp` by default as well.