[r=mpontillo][bug=][author=allenap] New view maasserver_routable_pairs.
This view selects pairs of IP addresses that can route between nodes. All addresses on a fabric are mutually routable so this view finds pairs of IP addresses that are on the same fabric. This view should typically be selected with additional constraints on, say, the sets of nodes, to find addresses that are mutually routable between region controllers for example.