[r=mpontillo][bug=][author=allenap] New --select-bucket option for the test scripts.
Tests can be split into a configurable number of buckets and then the tests for one of those buckets can be then run. This is a step towards being able to parallelise test runs. The mechanism for choosing buckets is stable so this will work locally or in a distributed environment.