[r=mpontillo][bug=][author=ltrager] Use ScriptSet and ScriptResult in place of NodeResult for commissioning and installation.
MAAS now creates a ScriptSet and an empty set of ScriptResults at the start of commissioning, controller refresh, and node deployment. By creating the entries at the start of each process users will soon be able to track the progress of each script being run on the node. As results are uploaded to MAAS the results are captured in the ScriptResults model, the NodeResults model is no longer used.