[r=blake-rouse][bug=][author=ltrager] Add the test API operation which boots the node into an ephemeral environment and runs tests.
Like the commissioning operation the testing_scripts option can be passed to specify which scripts should be run. If no option is passed test scripts tagged 'commissioning' will be run. Testing can be aborted which will cause the node to go back to the previous status it was in unless the previous status was commissioning. If the previous status was commissioning the status will be set to ready. This allows users to abort auto testing and have a usable node.