[r=newell-jensen][bug=][author=ltrager] Remove the test action from the UI.
When creating the test API operation I had to add a Test class to the node_actions list. This allows the TestForm to validate that the node is in a state which can transition to testing. This same list is used to generate the actions displayed in the UI. I removed Test from the main list so Test doesn't appear in the UI. The form now creates its own list of node actions to validate against. Once we have the full testing UI this can be removed.