[r=allenap][bug=][author=ltrager] Add smartctl-validate, smartctl-short, smartctl-long, and smartctl-conveyance tests.
Each test runs against all drives in the system in parallel. Since we're only reading SMART data, or asking the drive to run self-tests, this shouldn't saturate the bus. The full output of smartctl --xall is returned for all supported drives. smartctl-validate is configured to run by default when no testing_scripts option is passed to the commissioning or test operation.