[r=andreserl,mpontillo,allenap][bug=1664285,1665478][author=ltrager] Capture stdout, stderr, and the combined output of all scripts run.
maas-run-remote-scripts now captures stdout, stderr, and the combined output as three separate files. The metadataservice has been modified to allow uploading and storing all three files in the ScriptResult model. Uploaded files ending in .err will be stored as the ScriptResult's stderr output, files ending in .out will be stored as the ScriptResult's stdout output, all other files will be stored as the ScriptResult's combined output.

This also fixes 1664285 to return results base64 encoded and 1665478 to show the status as testing.