[r=andreserl,blake-rouse][bug=][author=ltrager] Add a ScriptForm
The ScriptForm validates all Script fields.  The script_type can be specified by numeric ID or name('testing' or 'commissioning'). If a comment is specified the comment is added to the VersionedTextFile. The form only allows the tags and timeout fields to be edited if the Script is a default script.