[r=mpontillo][bug=][author=mpontillo] Implement tests for Netplan YAML.Support netplan (v2 YAML) rendering in the network preseed generator.
Drive-by fix to allow preseed YAML to be rendered for nodes with NULL IP addresses (such as AUTO addresses that have not been assigned yet) to be rendered (without those addresses). It would be nice if a special syntax existed for "TBD" addresses, so we could show these for debugging purposes.