[r=mpontillo][bug=][author=ltrager] Enable the script reaper for commissioning nodes.
Commissioning nodes, like testing nodes, now have 20 minutes to boot and start commissioning or testing. Once commissioning or testing as begun the region checks that it has received a heartbeat from the node. If it has missed the last 5 heart beats(10 minutes) the node status is set to failed. The region also checks if any currently running script is over its time limit, if it has the running script and any pending scripts are set to timedout status and the node status is set to failed.