[r=mpontillo][bug=][author=ltrager] Validate script_sets before deletion.
Don't allow deleting current_commissioing_script_set as it is the data set MAAS used to gather hardware information about the node.

The current_installation_script_set is only set when a node is deployed. Don't allow it to be deleted as it contains info about the OS deployed.

MAAS uses the current_testing_script_set to know what testing script set should be shown by default in the UI. If an older version exists set the current_testing_script_set to it.