[r=blake-rouse][bug=1677056][author=ltrager] Add installation tab to UI
The installation tab appears when a node has a current_installation_script_set. When no output is available a status message now appears. Like commissioning and testing a status icon is on the installation tab. To capture when installation starts the metadata server now looks for the status start message from curtin.