[r=blake-rouse][bug=][author=newell-jensen] Add remote storage discovery and composition to RSD Pod driver.
Discovery finds the total available Logical Volume Groups and the largest Logical Volume in the pod and assigns them to the discovered pod and the discovered pod hints, respectively.  Several helper methods for scraping and calculating the remote storage are also included.

Composition add the ability to compose a node with remote storage capabilities.